app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
PyMuPDF==1.23.26
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==1.10.9
aiohttp==3.8.5
openai==0.27.8
sqlalchemy==1.4.50
alembic==1.11.1
python-dotenv==1.0.0
requests==2.31.0
